@@ -1,11 +1,7 @@
 import React from 'react';
 import clsx from 'clsx';
 import Link from '@docusaurus/Link';
-import {
-  findFirstCategoryLink,
-  useDocById,
-} from '@docusaurus/theme-common/internal';
-import isInternalUrl from '@docusaurus/isInternalUrl';
+import {useDocById} from '@docusaurus/theme-common/internal';
 import {translate} from '@docusaurus/Translate';
 import styles from './styles.module.css';
 function CardContainer({href, children}) {
@@ -34,7 +30,7 @@ function CardLayout({href, icon, title, description}) {
   );
 }
 function CardCategory({item}) {
-  const href = findFirstCategoryLink(item);
+  const href = item.href
   // Unexpected: categories that don't have a link have been filtered upfront
   if (!href) {
     return null;
